Added: figures to accompany issue #72
These files support issue #72 and are not intended to be merged with
master.

`unnamed-chunk-1.eps` and `unnamed-chunk-1.pdf` are referenced by `bug.md`
while `unnamed-chunk-1-eps-converted-to.pdf` was created by pandoc when
converting `bug.md` to `bug.pdf`. `bug.Rmd` is the MWE but requires ℝ
and knitr (an ℝ package) to build.